Make the topk_aggregate benchmark's make_data function public
